// src/app/models/login.metadata.ts
export interface LoginMetaData {
  IH_LoginPage_Header: string;
  IH_Google_Button: string;
  IH_Or: string;
  IH_Forgot_Password: string;
  IH_Login_Button: string;
  IH_Login_Footer: string;
  IH_Login_Footer_Sign_Up: string;
  IH_Login_Email: string;
  IH_Login_Password: string;
  IH_Login_Incorrect: string;
  IH_Login_Hi: string;
  IH_Login_Success_Message: string;
  IH_Login_Welcome: string;
  IH_LoginPage_App_Title: string;
  IH_LoginPage_Image_Text_1: string;
  IH_LoginPage_Image_Text_2: string;
  IH_LoginPage_Image_Text_3: string;
  IH_LoginPage_Email_Us: string;
  IH_LoginPage_Email_Blocked: string;
  IH_LoginPage_Cred_Error: string;
  IH_Login_LoginEmail: string;
  IH_Login_ContAppleID: string;
  IH_Login_UnauthorizedUser: string;
  IH_Login_Redirect: string;
}

export interface LoginExceptionMetadata {
  IH_LoginExceptionPage_App_Title: string;
  IH_LoginExceptionPage_Send_Email: string;
  IH_LoginExceptionPage_Verification_Pending: string;
  IH_LoginExceptionPage_Verify_Email: string;
}